Fixes #32534 - Move smart proxy dynflow core into smart proxy dynflow
