+++ lib/smart_proxy_dynflow/core.rb
@@ -1,4 +1,4 @@
-module SmartProxyDynflowCore
+class Proxy::Dynflow
   class Core
     attr_accessor :world, :accepted_cert_serial
