+++ lib/smart_proxy_dynflow/middleware/keep_current_request_id.rb
