+++ lib/smart_proxy_dynflow/task_launcher_registry.rb
@@ -1,0 +1,31 @@
+class Proxy::Dynflow
+  class TaskLauncherRegistry
+    class << self
+      def register(name, launcher)
+        registry[name] = launcher
+      end
+
+      def fetch(name, default = nil)
+        if default.nil?
+          registry.fetch(name)
+        else
+          registry.fetch(name, default)
+        end
+      end
+
+      def key?(name)
+        registry.key?(name)
+      end
+
+      def operations
+        registry.keys
+      end
+
+      private
+
+      def registry
+        @registry ||= {}
+      end
+    end
+  end
+end
