+++ lib/smart_proxy_dynflow_core.rb
@@ -1,17 +1,12 @@
-require 'dynflow'
-require 'smart_proxy_dynflow_core/middleware/keep_current_request_id'
+# Whatever is here is just a compatibility layer
+# Once all the _core have been migrated, this can be dropped.
+
+require 'smart_proxy_dynflow'
+
+# REX core explicitly requires this file, otherwise we could use the trick we
+# use with callback
 require 'smart_proxy_dynflow_core/task_launcher_registry'
-require 'foreman_tasks_core'
-require 'smart_proxy_dynflow_core/log'
-require 'smart_proxy_dynflow_core/settings'
-require 'smart_proxy_dynflow_core/core'
-require 'smart_proxy_dynflow_core/helpers'
-require 'smart_proxy_dynflow_core/callback'
-require 'smart_proxy_dynflow_core/api'
 
 module SmartProxyDynflowCore
-  Core.after_initialize do |dynflow_core|
-    ForemanTasksCore.dynflow_setup(dynflow_core.world)
-  end
-  Core.register_silencer_matchers ForemanTasksCore.silent_dead_letter_matchers
+  Callback = Proxy::Dynflow::Callback
 end
